Growatt - Refactored sensor types
